BUG: Correct expected values in various tests
Some tests were leaking changes to some of the static
inputs. All Empyrical functions are now safe from
mutating their inputs, however some of the sample
values generated for the tests expected the mutated
Series.

In most cases, the old value can be generated by
replacing returns[0] with 0.01 or -0.01 (depending
on the order of the tests).

In addition, this exposed two calculations which
were not nan-tolerant. The tests using mixed_returns
are expected to evaluate the nan-tolerance of each
calculation, however the cross-test state changes
to mixed_returns was pre-empting that. This is
also fixed.